requirements.txt
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.3
djangorestframework==3.14.0
minio==7.1.15
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
